+30 mins Created Game Manager, Game, Game Settings
Created a game manager to handle the management of runtime game data such as loading new games. Created Game obj to hold the state of the given game. Created Game Settings obj to handle the runtime settings of the game.

Functionality will be added to these scripts when more specific settings are created.