Core Logic Implementation
Get various classes and game objects working together to create all necessary components for a game. Season Manager has an Initialize State & Start State. You provide a season manager with the number of players you want to create a season for. It then instantiates the number of players needed in the Player Pool Manager.

Starting the Season will then initiate the Round Manager. This looks at the number of players and the opponent type for the round (UnRanked, Ranked, Head to Head) and generates the number of games needed to play that round.  It then asks the Player Pool Manager to pool the players needed into the various games, generating CPU Players as needed.