Scoring Logic Completed
Implemented the scoring logic. Hands can now be passed into the score manager and it will evaluate what type of hand it is (e.g. Straight, Two Pair, etc.)  It will also give the hand a high score in case of a tie (e.g., Pair of 3s v Par of 4s)